+++ Options.lua
@@ -3254,6 +3254,13 @@ local function CreateAuraAreaLayoutOptions(pos, widget_info)
       values[widget_info] = nil
       return values
     end,
+    set = function(info, val)
+      if val ~= "Healthbar" and db.AuraWidget[val].AnchorTo == widget_info then
+        Addon.Logging.Error(L["Cyclic anchoring of aura areas to each other is not possible."], string.format(L["%s already anchored to %s"], val, widget_info))
+      else
+        SetValueWidget(info, val)
+      end
+    end,
     arg = { "AuraWidget", widget_info, "AnchorTo" }
   }