+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # @project-version@ (@build-time@)
 
 * Fixed a bug with CCs not being shown when debuffs are not enabled [Comment #5983].
+* Fixed a Lua error when aura areas where anchored to each other in a cyclic way [Issue #520].